+++ src/constants/messages.ts
@@ -4,29 +4,34 @@ export const workspaceNotTrusted = 'Workspace is not trusted';
 export const gitRepoNotFound =
   "Git repository not found. Try run 'git init' before using";
 
-export const cannotReadContent = gitRepoNotFound;
+export const cannotReadContent = 'Cannot read changelists from file';
 
 export const cannotWriteContent = 'Cannot write changelists to file';
 
-export const askToInitExtFiles = 'Would you like to initialize Git Changelists ? \nYou can do it later using command "Initialize Git Changelists"';
+export const askToInitExtFiles =
+  'Would you like to initialize Git Changelists ? \nYou can do it later using command "Initialize Git Changelists"';
 export enum AskToInitAnswers {
   yes = 'Yes',
   no = 'No, later',
-};
+}
 
 export const initializingExtFiles = 'Initializing git-changelists!';
 
-export const changelistNameMandatory = 'A Changelist name is mandatory to execute this action';
+export const changelistNameMandatory =
+  'A Changelist name is mandatory to execute this action';
 
-export const changelistNameAlreadyExists = 'Changelist with such name already exists';
+export const changelistNameAlreadyExists =
+  'Changelist with such name already exists';
 
 export const changelistNotFound = 'Changelist not found';
 
-export const fileWasRestored = '{file} is restored from the state assumed to be unchanged.';
+export const fileWasRestored =
+  '{file} is restored from the state assumed to be unchanged.';
 
 export const fileAssumedUnchanged = '{file} is assumed to be unchanged.';
 
-export const selectChagelistToAddFile = 'Select Changelist where you want to add file';
+export const selectChagelistToAddFile =
+  'Select Changelist where you want to add file';
 
 export const newChangelistPlaceholder = 'New Changelist name';